Make AckV4 behave the same as AuthV4
The AckRespV4 struct is slightly different than the AuthMsgV4 struct.
The latter manipulates the keys when serializing and deserializing,
turning them from raw to der and viceversa.

This subtle difference presents a problem in other places when we expect
those keys to be of similar type (either raw or der). For example, the
`Secrets.derive_secrets` can be used when receiving and auth message or
an ack message.

For those reason, we make AckRespV4 behave like AuthMsgV4, making the
key raw when serializing to be sent to our peers, and making it der when
deserializing the data received from a peer.