WIP Test handshake in RLPx test
We add a test for RLPx module that tests a handshake when a remote
initiates the auth. This test needs some cleaning up (lots of comments
left) and more importantly, it is failing intermittently. I do not know
why that is.

My first thought is that the nonce and the public keys are generated on
the fly. Perhaps some values do not work for the handshake.

Also added a little note on building and ack response in `Handshake`.
The fields in that struct (and in the `AuthMsgV4` struct) are confusing
because they are all called `remote_*`. It would be better to follow the
convention outlined in
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-8.md of calling
them `initiator_*` and `recipient_*`.